backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-generativeai
pygithub
python-dotenv
requests